requirements.txt
ollama
python-dotenv
rich
pytest
ruff